requirements.txt
alembic==0.8.10
appdirs==1.4.0
bcrypt==3.1.4
cffi==1.11.2
click==6.7
configparser==3.5.0
decorator==4.0.11
et-xmlfile==1.0.1
Flask==0.12
Flask-Login==0.4.1
flask-markdown2==0.0.3
Flask-Migrate==2.0.2
Flask-MySQL==1.4.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
gunicorn==20.0.4
itsdangerous==0.24
jdcal==1.3
Jinja2==2.9.4
jinja2-markdown==0.0.3
legacy==0.1.6
loginform==1.2.0
lxml==4.4.2
Mako==1.0.6
Markdown==2.6.9
MarkupSafe==0.23
mysql-connector-python==8.0.18
mysqlclient==1.4.6
olefile==0.44
openpyxl==2.4.8
packaging==16.8
pbr==1.10.0
Pillow==7.0.0
protobuf==3.11.2
pycparser==2.18
PyMySQL==0.7.9
pyparsing==2.1.10
python-editor==1.0.3
PyYAML==5.3.1
six==1.10.0
SQLAlchemy==1.3.12
sqlalchemy-migrate==0.10.0
sqlalchemy-mptt==0.2.2
sqlparse==0.2.2
Tempita==0.5.2
Werkzeug==0.11.15
WTForms==2.1
